+++ packages/cortex-react/docs/DatePickerInput.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 import React, { useState } from 'react';
-import { DatePickerInput, DateRangePickerInput } from '../src';
+import { DatePickerInput, DateRange, DateRangePickerInput } from '../src';
 
 export default {
   title: 'Cortex/DatePickerInput',
@@ -9,34 +9,45 @@ export default {
 } as Meta<typeof DatePickerInput>;
 
 export const Default: StoryObj<typeof DatePickerInput> = {
-  render: () => {
-    const [value, setValue] = useState(new Date());
+  args: {
+    label: 'Select Date',
+    disabled: false,
+  },
+  render: args => {
+    const [value, setValue] = useState<Date | undefined>(new Date());
 
     return (
       <div className={'w-[200px] h-[450px]'}>
         <DatePickerInput
           value={value}
           onChange={value => setValue(value)}
-          label="Select date"
+          label={args.label}
+          disabled={args.disabled}
         />
       </div>
     );
   },
 };
 
+const start = new Date();
+const end = new Date();
+end.setDate(end.getDate() + 2);
+
 export const Range: StoryObj<typeof DateRangePickerInput> = {
-  render: () => {
-    const start = new Date();
-    const end = new Date();
-    end.setDate(end.getDate() + 2);
-    const [value, setValue] = useState({ start, end });
+  args: {
+    label: 'Select date range',
+    disabled: false,
+  },
+  render: args => {
+    const [value, setValue] = useState<DateRange | undefined>({ start, end });
 
     return (
       <div className={'w-[250px] h-[450px]'}>
         <DateRangePickerInput
           value={value}
           onChange={value => setValue(value)}
-          label="Select date range"
+          label={args.label}
+          disabled={args.disabled}
         />
       </div>
     );