+++ packages/cortex-react/src/components/DatePicker/DatePickerInput.tsx
@@ -1,42 +1,41 @@
-import { CalendarDate } from '@internationalized/date';
 import React from 'react';
 import { useDatePickerInput } from '../../hooks';
 import { dateToCalendarDate } from '../../utils';
 import { Calendar } from '../Calendar/Calendar';
-import { InputPropsBase } from '../Input';
+import { InputProps } from '../Input';
 import { Popover } from '../Popover';
 import { usePopoverContext } from '../Popover/Context';
-import { Content } from './Content';
+import { Content } from '../Select/Content';
 import { DateField } from './DateField';
 import { DatePickerInputBase } from './DatePickerInputBase';
 
-export interface DatePickerInputProps extends InputPropsBase {
+export interface DatePickerInputProps
+  extends Omit<InputProps, 'value' | 'onChange'> {
   value?: Date;
-  onChange: (date: Date) => void;
+  onChange: (date?: Date) => void;
 }
 
 const DatePickerInputWithPopover = (props: DatePickerInputProps) => {
   const { setIsOpen } = usePopoverContext();
-  const { onChange, value, label, variants } = props;
+  const { onChange, value, label, variants, disabled } = props;
   const { fieldProps, state, ref } = useDatePickerInput({ value, onChange });
 
   return (
     <div ref={ref} data-testid="date-picker-input">
-      <Popover.Trigger>
+      <Popover.Trigger disabled={true}>
         <DatePickerInputBase
           variants={{
             ...variants,
             intent: state.isInvalid ? 'error' : variants?.intent,
           }}
           label={label}
+          disabled={disabled}
+          value={value}
+          onClean={() => state.setValue(null)}
+          onToggle={() => setIsOpen(open => !open)}
         >
           <div>
-            <DateField
-              {...fieldProps}
-              onChange={value => {
-                state.setDateValue(value as CalendarDate);
-              }}
-            />
+            <DateField {...fieldProps} isDisabled={disabled} />
           </div>
         </DatePickerInputBase>
       </Popover.Trigger>
@@ -52,7 +51,8 @@ const DatePickerInputWithPopover = (props: DatePickerInputProps) => {
     </div>
   );
 };
-/** DatePickerInput component */
+
+/** DatePickerInput component, extends HTML Input Element props*/
 export const DatePickerInput = (props: DatePickerInputProps) => (
   <Popover.Root placement="bottom-start" trigger="click">
     <Content>