+++ packages/cortex-react/src/components/Popover/Trigger.tsx
@@ -5,11 +5,26 @@ import { usePopoverContext } from './Context';
 export interface PopoverTriggerProps {
   /** child element */
   children: React.ReactElement;
+  disabled?: boolean;
 }
 
-export const PopoverTrigger = ({ children }: PopoverTriggerProps) => {
-  const { triggerProps } = usePopoverContext();
-  const isDisabled = children.props?.disabled;
+const isDisabledProps = {
+  onClick: () => undefined,
+  onKeyDown: () => undefined,
+  onMouseDown: () => undefined,
+  onKeyUp: () => undefined,
+  onPointerDown: () => undefined,
+};
 
-  return cloneWithProps(children, isDisabled ? {} : triggerProps);
+export const PopoverTrigger = ({
+  children,
+  disabled = false,
+}: PopoverTriggerProps) => {
+  const { triggerProps } = usePopoverContext();
+  const isDisabled = disabled ?? children.props?.disabled;
+
+  return cloneWithProps(children, {
+    ...triggerProps,
+    ...(isDisabled ? isDisabledProps : {}),
+  });
 };