+++ packages/cortex-react/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.12-beta.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-react@1.9.12-beta.0...@tecsinapse/cortex-react@1.9.12-beta.1) (2024-12-26)
+
+### Bug Fixes
+
+- disabled on datepickerinput icons + tests. ([a40736a](https://github.com/tecsinapse/design-system/commit/a40736ac142e263dfabfcfcba99b825d37a41004))
+- remove double onToggle ([6e236d5](https://github.com/tecsinapse/design-system/commit/6e236d5018dd49642c018072c10f06f86f48c882))
+
 ## [1.9.12](https://github.com/tecsinapse/design-system/compare/@tecsinapse/cortex-react@1.9.12-beta.0...@tecsinapse/cortex-react@1.9.12) (2024-12-18)
 
 **Note:** Version bump only for package @tecsinapse/cortex-react